# docs/api.yaml
openapi: 3.0.3
info:
  title: Auction Service API
  version: "1.0.0"
  description: |
    OpenAPI specification for the auction platform REST API.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
paths:
  /api/bidders/onboarding:
    post:
      summary: Onboard a bidder
      description: Requires `admin.write` scope.
      operationId: onboardBidder
      tags: [Bidders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardBidderRequest'
      responses:
        '201':
          description: Bidder onboarded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardBidderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/bidders/{bidderId}:
    put:
      summary: Update a bidder
      description: Requires `admin.write` scope.
      operationId: updateBidder
      tags: [Bidders]
      parameters:
        - $ref: '#/components/parameters/BidderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBidderRequest'
      responses:
        '200':
          description: Bidder updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardBidderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      summary: Soft delete a bidder
      description: Requires `admin.write` scope.
      operationId: deleteBidder
      tags: [Bidders]
      parameters:
        - $ref: '#/components/parameters/BidderId'
      responses:
        '204':
          description: Bidder soft-deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/bidders/{bidderId}/block:
    post:
      summary: Block a bidder
      description: Blocks bidder for configured duration.
      operationId: blockBidder
      tags: [Bidders]
      parameters:
        - $ref: '#/components/parameters/BidderId'
      responses:
        '204':
          description: Bidder blocked
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/bidders/{bidderId}/unblock:
    post:
      summary: Manually unblock a bidder
      description: Requires `admin.write` scope.
      operationId: unblockBidder
      tags: [Bidders]
      parameters:
        - $ref: '#/components/parameters/BidderId'
      responses:
        '204':
          description: Bidder unblocked
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auctions:
    post:
      summary: Create an auction
      description: Creates an auction in `DRAFT` state.
      operationId: createAuction
      tags: [Auctions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuctionRequest'
      responses:
        '201':
          description: Auction created
          headers:
            Location:
              description: URI of the created auction
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAuctionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auctions/{auctionId}/start:
    post:
      summary: Start an auction
      operationId: startAuction
      tags: [Auctions]
      parameters:
        - $ref: '#/components/parameters/AuctionId'
      responses:
        '204':
          description: Auction started
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/auctions/{auctionId}/schedule:
    post:
      summary: Schedule an auction
      operationId: scheduleAuction
      tags: [Auctions]
      parameters:
        - $ref: '#/components/parameters/AuctionId'
      responses:
        '204':
          description: Auction scheduled
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/auctions/{auctionId}/close:
    post:
      summary: Close an auction
      operationId: closeAuction
      tags: [Auctions]
      parameters:
        - $ref: '#/components/parameters/AuctionId'
      responses:
        '204':
          description: Auction closed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/auctions/{auctionId}/cancel:
    post:
      summary: Cancel an auction
      operationId: cancelAuction
      tags: [Auctions]
      parameters:
        - $ref: '#/components/parameters/AuctionId'
      responses:
        '204':
          description: Auction cancelled
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/auctions/{auctionId}/settle:
    post:
      summary: Settle an auction
      operationId: settleAuction
      tags: [Auctions]
      parameters:
        - $ref: '#/components/parameters/AuctionId'
      responses:
        '204':
          description: Auction settled
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/auctions/{auctionId}/bids:
    post:
      summary: Place a bid
      description: Requires `bid.write` scope. `idempotencyKey` must be unique per `bidderId`.
      operationId: placeBid
      tags: [Bids]
      parameters:
        - $ref: '#/components/parameters/AuctionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceBidRequest'
      responses:
        '202':
          description: Bid accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceBidResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Bid rejected by business rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceBidResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AuctionId:
      name: auctionId
      in: path
      required: true
      description: Auction identifier
      schema:
        type: string
        format: uuid
    BidderId:
      name: bidderId
      in: path
      required: true
      description: Bidder identifier
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request payload or business validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid JWT
    Conflict:
      description: Business state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CreateAuctionRequest:
      type: object
      required: [title, description, reservePrice, minIncrement, startTime, endTime]
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 4000
        reservePrice:
          type: number
          format: double
          minimum: 0.01
        minIncrement:
          type: number
          format: double
          minimum: 0.01
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    CreateAuctionResponse:
      type: object
      required: [auctionId]
      properties:
        auctionId:
          type: string
          format: uuid
    OnboardBidderRequest:
      type: object
      required: [firstName, lastName, email, nationalId, purchasingAuthorizationLimit]
      properties:
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 255
        nationalId:
          type: string
          maxLength: 50
        purchasingAuthorizationLimit:
          type: number
          format: double
          minimum: 0.01
    UpdateBidderRequest:
      type: object
      required: [firstName, lastName, email, nationalId, purchasingAuthorizationLimit]
      properties:
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 255
        nationalId:
          type: string
          maxLength: 50
        purchasingAuthorizationLimit:
          type: number
          format: double
          minimum: 0.01
    OnboardBidderResponse:
      type: object
      required: [bidderId, firstName, lastName, email, nationalId, purchasingAuthorizationLimit]
      properties:
        bidderId:
          type: string
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 255
        nationalId:
          type: string
          maxLength: 50
        purchasingAuthorizationLimit:
          type: number
          format: double
          minimum: 0.01
        blockedUntil:
          type: string
          format: date-time
          nullable: true
    PlaceBidRequest:
      type: object
      required: [bidderId, amount, idempotencyKey]
      properties:
        bidderId:
          type: string
        amount:
          type: number
          format: double
          minimum: 0.01
        idempotencyKey:
          type: string
          description: Client-generated idempotency token; must be unique per bidder.
    PlaceBidResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ACCEPTED, REJECTED]
        rejectReason:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string